Miner address retrieval checked

+++ common/script/creation.sql
@@ -56,10 +56,10 @@ CREATE TABLE place_of_interest (
 
 DROP TABLE IF EXISTS poi_raw_geo;
 CREATE TABLE poi_raw_geo (
-       poi_id INT,
+       poi_id  INT,
        address VARCHAR(300),
-       lat     FLOAT,
-       lng     FLOAT,
+       lat     DECIMAL(11, 8),
+       lng     DECIMAL(11, 8),
 
        FOREIGN KEY (poi_id) REFERENCES place_of_interest(id) ON UPDATE RESTRICT ON DELETE RESTRICT
 ) DEFAULT CHARACTER SET=utf8;
